Reorganise project to use pkg structure for plugins and add deps to go module
