+++ go.mod
@@ -1,3 +1,5 @@
 module gitub.com/stryda/prompt
 
 go 1.13
+
+require gopkg.in/src-d/go-git.v4 v4.13.1
