+++ pkg/directory/directory.go
@@ -1,0 +1,7 @@
+package directory
+
+import "path/filepath"
+
+func Print(path string) string {
+	return filepath.Base(path)
+}
