--- plugins/directory.go
@@ -1,7 +1,0 @@
-package plugins
-
-import "path/filepath"
-
-func PrintDirectory(path string) string {
-	return filepath.Base(path)
-}
